feat: do not use Gem module
When using Ruby package managers, they will not be picked up by the
Gem module, resulting in the gems being installed in the system path,
which is not desirable in most cases.

Using the Shell module with a login shell will work since it sources
the usual *profile files, though idempotency must be ensured manually.