feat: improve Xcode install
Fall back to manual download since xcodes can no longer do it without
signing in as Apple patched an apparent loophole. A valid URL to a XIP
must be provided. No such thing as a free lunch.

In order to accommodate Xcode 15 packaging format, this role now also
installs the iOS runtime corresponding to an Xcode version. However,
this mapping must be set in the respective `xcodes` variable object.

Read the comments for other caveats and more information.